chore: 이미지 업로드 api와 FileEdit 컴포넌트 공통 폴더로 이동
