+++ src/api/index.ts
@@ -1,2 +1,3 @@
 export { default as http } from './core';
+export { default as postImage } from './api';
 export * from './types';
