+++ src/app/mypage/[blogId]/components/Common/index.tsx
@@ -1,3 +1,2 @@
 export { default as FileThumbnails } from './FileThumbnails';
 export { default as useReadFile } from './useReadFile';
-export { default as FileEdit } from './FileEdit';
