+++ src/components/Common/index.tsx
@@ -11,3 +11,4 @@ export { default as ToastProvider } from './Toast';
 export { default as TextArea } from './TextArea';
 export { default as Pagination } from './Pagination';
 export { default as Dropdown } from './Dropdown';
+export { default as FileEdit } from './FileEdit';
